Move attachment check in build()

+++ src/pushover/constants/pushover_consts.rs
@@ -2,4 +2,4 @@
 
 pub const PUSHOVER_API_ENDPOINT: &str = "https://api.pushover.net/1/messages.json";
 pub const PUSHOVER_API_DOCUMENTATION_URL: &str = "https://pushover.net/api";
-pub const PUSHOVER_API_ATTACHMENT_MAX_SIZE_BYTES: u32 = 2621440;
+pub const PUSHOVER_API_ATTACHMENT_MAX_SIZE_BYTES: u64 = 2621440;
